Separated factors from base
