+++ heating_planner/app.py
@@ -3,7 +3,7 @@ import streamlit_authenticator as stauth
 import yaml
 from heating_planner.front.tabs import map, parameters, reference_definition, raw_data
 
-st.set_page_config(layout="wide")
+st.set_page_config(layout="wide", page_title="Heating Planner", page_icon="🌍")
 
 with open("./credentials.yaml") as f:
     config = yaml.load(f, Loader=yaml.loader.SafeLoader)
