+++ heating_planner/back/data/base.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from enum import StrEnum
 from typing import Dict, List
 from loguru import logger
 import geopandas as gpd
@@ -7,64 +6,14 @@ import pandas as pd
 from shapely.geometry import Point
 import os
 import tempfile
-import shutil
 from urllib.parse import urlparse
 import requests
 
 from heating_planner.back.geo.coder import geocoding
+from heating_planner.back.data.model.factor import Factor, FactorType
 from heating_planner.crs import METRIC_CRS
 
 SJOIN_MAX_DISTANCE_M = 8_000  # meters
-
-
-class FactorTrend(StrEnum):
-    HIGHER_BETTER = "higher_better"
-    LOWER_BETTER = "lower_better"
-    NEUTRAL = "neutral"
-
-    @classmethod
-    def from_string(cls, value: str) -> "FactorTrend":
-        """Convert string to FactorType enum value"""
-        try:
-            return cls(value.lower())
-        except ValueError:
-            raise ValueError(f"Invalid FactorType: {value}. Must be one of {[t.value for t in cls]}")
-
-
-class FactorType(StrEnum):
-    CONTINUOUS = "continuous"
-    DISCRETE = "discrete"
-    BINARY = "binary"
-
-
-@dataclass
-class Factor:
-    name: str
-    description: str
-    trend: FactorTrend
-    type: FactorType
-    unit: str
-
-    def __hash__(self):
-        return hash(self.name)
-
-    def __repr__(self):
-        return f"{self.name} ({self.type}): {self.description[:50]}... "
-
-    def is_binary(self):
-        return self.type == FactorType.BINARY
-
-    def is_continuous(self):
-        return self.type == FactorType.CONTINUOUS
-
-    def __eq__(self, other):
-        return self.name == other.name
-
-    def __lt__(self, other):
-        return self.name < other.name
-
-    def __le__(self, other):
-        return self.name <= other.name
 
 
 @dataclass
