+++ heating_planner/back/data/real_estate.py
@@ -2,7 +2,8 @@ from typing import List
 import geopandas as gpd
 import pandas as pd
 
-from heating_planner.back.data.base import METRIC_CRS, Factor, FactorTrend, FactorType, HazardDataset
+from heating_planner.back.data.base import HazardDataset, METRIC_CRS
+from heating_planner.back.data.model.factor import Factor, FactorTrend, FactorType
 
 
 FACTOR_NAME = "real_estate_prices"
