+++ heating_planner/back/data/sea.py
@@ -1,6 +1,8 @@
 import geopandas as gpd
 from loguru import logger
-from heating_planner.back.data.base import Factor, FactorTrend, FactorType, HazardDataset, METRIC_CRS
+
+from heating_planner.back.data.base import HazardDataset, METRIC_CRS
+from heating_planner.back.data.model.factor import Factor, FactorTrend, FactorType
 
 FACTOR_NAME = "sea_1m_sub"
 FACTOR_DESCR = "Zones innondées avec 1m d'élévation"
